Removes the device sensor class from HCHO (#42)
This is because of a compatibility problem between units for HCHO and
the previously selected device sensor class for VOC. HCHO is measured in
mg/m3 while VOC is measured in µg/m3. This caused Home Assistant to drop
a warning in the logs every time a new Dyson Formaldehyde device was
added.

So now we're not using any sensor class for HCHO, which defaults it to
None and allows us to use arbitrary units.